feat: rename to i18n-easy-edit
