+++ shell/app/i18n-easy-edit/pages/edit-modal/index.tsx
