+++ shell/app/i18n-easy-edit/pages/i18n-wrapper/css/index.less
