+++ shell/app/i18n-easy-edit/pages/index.tsx
